requirements.txt
colorlog~=4.4.0
dataclasses~=0.6
dill~=0.3.2
gitpython
matplotlib~=3.2.1
numpy~=1.18.4
pillow>=8.2.0
prompt_toolkit~=3.0.5
PyYAML==5.4
scikit-learn~=0.23.2
scipy~=1.5.4
sklearn~=0.0
tensorboard~=2.3.0
#torch~=1.7.0
#torchvision~=0.8.1
transformers~=2.1.1
#torchtext~=0.7.0
tqdm~=4.46.0
